% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{as_brmsfit}
\alias{as_brmsfit}
\title{Extract the brmsfit from a Bayesian uncounted Model}
\usage{
as_brmsfit(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
The underlying \code{brmsfit}.
}
\description{
Extract the brmsfit from a Bayesian uncounted Model
}
