% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich_methods.R
\name{estfun.uncounted}
\alias{estfun.uncounted}
\title{Empirical estimating functions (score contributions)}
\usage{
\method{estfun}{uncounted}(x, ...)
}
\arguments{
\item{x}{An uncounted object}

\item{...}{Ignored}
}
\description{
Returns an n x p matrix of per-observation score contributions
for use with the sandwich package. Row i is the gradient of the
(quasi-)log-likelihood contribution of observation i with respect
to the parameter vector.
}
