% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich_methods.R
\name{hatvalues.uncounted}
\alias{hatvalues.uncounted}
\title{Hat values (leverage)}
\usage{
\method{hatvalues}{uncounted}(model, ...)
}
\arguments{
\item{model}{An uncounted object}

\item{...}{Ignored}
}
\description{
Computes diagonal of the hat matrix
\eqn{H = W^{1/2} Z (Z'WZ)^{-1} Z' W^{1/2}}
where Z is the model matrix and W is the diagonal weight matrix.
}
