% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich_methods.R
\name{model.matrix.uncounted}
\alias{model.matrix.uncounted}
\title{Extract model matrix}
\usage{
\method{model.matrix}{uncounted}(object, ...)
}
\arguments{
\item{object}{An uncounted object}

\item{...}{Ignored}
}
\description{
Returns the Jacobian/design matrix used in sandwich computation.
When gamma was estimated, the returned matrix includes the gamma column.
}
