% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{nobs.uncounted}
\alias{nobs.uncounted}
\title{Number of observations for uncounted models}
\usage{
\method{nobs}{uncounted}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"uncounted"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
An integer scalar.
}
\description{
Returns the number of observations used to fit the model.
}
