% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_estimate.R
\name{nonlinear_estimate}
\alias{nonlinear_estimate}
\title{The Zhang (2008) model}
\usage{
nonlinear_estimate(data, counts, pop_size, auxiliary, family, control)
}
\arguments{
\item{data}{input data.frame}

\item{counts}{observed counts}

\item{pop_size}{population size}

\item{auxiliary}{auxiliary variable (single)}

\item{family}{what to assume: gaussian, poisson, NB}

\item{control}{control parameters (TODO)}
}
\description{
The Zhang (2008) model
}
