% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot.hidden}
\alias{plot.hidden}
\title{Plot diagnostics for a \code{hidden} object}
\usage{
\method{plot}{hidden}(x, which = NULL, label = FALSE)
}
\arguments{
\item{x}{an object of class \code{hidden} returned by the main estimation function \code{estimate_hidden_pop()}}

\item{which}{optional integer indication which plot should be displayed. If \code{NULL}, the default set of plots is shown.}

\item{label}{logical. If \code{TRUE} adds observation labels (country names) to the plots (only if country information is in the object).}
}
\description{
Diagnostic plotting method for objects of class \code{hidden}.
Depending on the estimation method used in \code{estimate_hidden_pop()} function, different sets of diagnostic
plots are displayed.
}
\details{
The function displays model diagnostics depending on the estimation method used.
The numbers in parentheses indicate the value to use in the \code{which} argument.
\itemize{
 \item \code{'ols'}: Residuals vs fitted (1), Q-Q plot (2), scale-location plot (3), residuals vs leverage (4), observed vs fitted (5).
 \item \code{'nls'}: Residuals vs fitted (1), Q-Q plot (2), observed vs fitted (3).
 \item \code{'glm'}: The same set of plots as in \code{ols}, but based on deviance/standardized deviance residuals.
 \item \code{'mle'}: √Observed vs √fitted (1), Q-Q plot for Anscombe residuals (2), absolute Anscombe residuals vs √Fitted (3), Anscombe residuals vs √Fitted (4).
}
If \code{which} is specified, only the corresponding plot is shown.
}
\examples{
\dontrun{
# Load Polish irregular migration data
data(foreigners_pl)

model_data <- subset(foreigners_pl, year == 2018 & half == 1)

# Basic Zhang model estimation using GLM (recommended)
result_zhang <- estimate_hidden_pop(
  data = model_data,
  observed = ~ border,
  auxiliary = ~ police,
  reference_pop = ~ pesel,
  method = "mle",
  family = "poisson"
)

plot(result_zhang)

}
}
