% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_hidden.R
\name{print.summary.hidden}
\alias{print.summary.hidden}
\title{Print method for summary.hidden objects}
\usage{
\method{print}{summary.hidden}(x)
}
\arguments{
\item{x}{An object of class \code{summary.hidden} returned by \code{summary.hidden()}.}
}
\description{
Print method for the \code{summary.hidden} object.
}
\details{
The function displays a formatted summary of the estimation results.
The output content depends on the estimation method used
in the \code{estimate_hidden_pop()} function and subsequently on the structure created by \code{summary.hidden}.
}
\examples{
\dontrun{
# Load Polish irregular migration data
data(foreigners_pl)

model_data <- subset(foreigners_pl, year == 2018 & half == 1)

# Basic Zhang model estimation using GLM (recommended)
result_zhang <- estimate_hidden_pop(
  data = model_data,
  observed = ~ border,
  auxiliary = ~ police,
  reference_pop = ~ pesel,
  method = "mle",
  family = "poisson"
)

s <- summary(result_zhang)
print(s)

}

}
