% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{residuals.uncounted_bayes}
\alias{residuals.uncounted_bayes}
\title{Bayesian Residuals}
\usage{
\method{residuals}{uncounted_bayes}(
  object,
  type = c("response", "pearson"),
  summary = c("mean", "median"),
  ...
)
}
\arguments{
\item{object}{An \code{"uncounted_bayes"} object.}

\item{type}{Residual type: response or Pearson.}

\item{summary}{Posterior summary used for fitted values.}

\item{...}{Additional arguments passed to \code{\link{fitted.uncounted_bayes}}.}
}
\value{
A numeric vector of residuals.
}
\description{
Bayesian Residuals
}
