% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_hidden.R
\name{summary.hidden}
\alias{summary.hidden}
\title{Summary method for class 'hidden'}
\usage{
\method{summary}{hidden}(object)
}
\arguments{
\item{object}{an object of class \code{hidden} returned by the main estimation function \code{estimate_hidden_pop()}}
}
\value{
An object of class \code{summary.hidden}, which is a list containing:
\itemize{
\item \code{method} - estimation method used
\item \code{coefficients} - estimation coefficients
\item \code{xi_est} - estimate of the target parameter
\item \code{conf_int_xi}, \code{conf_int_alpha}, \code{conf_int_beta} - confidence intervals
\item \code{vcov}, \code{vcov_method} - estimated variance-covariance matrix and method used
\item \code{se_coef} - standard errors for coefficients
\item \code{se_xi} - standard error for target parameter
\item \code{summary stat} - list of model-specific statistics
\item \code{aic}, \code{bic} - model selection criteria values
\item \code{residuals}, \code{fitted}, \code{m}, \code{n}, \code{N} - fitted values, residuals and observed data
}
}
\description{
The function constructs a summary list for objects of class \code{hidden}.
}
\details{
This function is typically used to prepare a structured summary that can then be printed using \code{print.summary.hidden()}.
Depending on the chosen estimation method in \code{estimate_hidden_pop()} function the following model statistics are calculated:
\itemize{
\item \strong{'ols'}: residual standard error, degrees of freedom, R-squared, adjusted R-squared, F-statistic and its p-value;
\item \strong{'nls'}: residual standard error, degrees of freedom, number of iterations, convergence status;
\item \strong{'glm'} - Poisson: null deviance, residual deviance, degrees of freedom, number of Fisher scoring iterations;
\item \strong{'mle'} - Zhang model: convergence status.
}
Other elements such as estimated coefficients, confidence intervals, standard errors, variance-covariance matrix, AIC, BIC, fitted values and residuals
are also included, regardless of the estimation method.
}
\examples{
\dontrun{
# Load Polish irregular migration data
data(foreigners_pl)

model_data <- subset(foreigners_pl, year == 2018 & half == 1)

# Basic Zhang model estimation using GLM (recommended)
result_zhang <- estimate_hidden_pop(
  data = model_data,
  observed = ~ border,
  auxiliary = ~ police,
  reference_pop = ~ pesel,
  method = "mle",
  family = "poisson"
)

summary(result_zhang)

}

}
