% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich_methods.R
\name{vcovCL.uncounted}
\alias{vcovCL.uncounted}
\title{Cluster-robust covariance for uncounted models}
\usage{
vcovCL.uncounted(
  x,
  cluster = NULL,
  type = NULL,
  sandwich = TRUE,
  fix = FALSE,
  ...
)
}
\arguments{
\item{x}{An uncounted object}

\item{cluster}{Clustering variable}

\item{type}{HC type}

\item{sandwich}{Return the full sandwich when \code{TRUE}}

\item{fix}{Passed to \code{sandwich::meatCL()}}

\item{...}{Ignored}
}
\description{
Cluster-robust covariance for uncounted models
}
