% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich_methods.R
\name{vcovHC.uncounted}
\alias{vcovHC.uncounted}
\title{Heteroskedasticity-consistent covariance for uncounted models}
\usage{
\method{vcovHC}{uncounted}(
  x,
  type = c("HC3", "const", "HC", "HC0", "HC1", "HC2", "HC4", "HC4m", "HC5"),
  omega = NULL,
  ...
)
}
\arguments{
\item{x}{An uncounted object}

\item{type}{HC type}

\item{omega}{Optional omega function passed to \code{sandwich::meatHC()}}

\item{...}{Ignored}
}
\description{
Heteroskedasticity-consistent covariance for uncounted models
}
