% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_estimate.R
\name{nonlinear_estimate}
\alias{nonlinear_estimate}
\title{Zhang (2008) model}
\usage{
nonlinear_estimate(data, counts, pop_size, auxiliary, backend, control)
}
\arguments{
\item{data}{input data.frame}

\item{counts}{observed counts}

\item{pop_size}{population size}

\item{auxiliary}{auxiliary variable (single)}

\item{backend}{whether to use OLS, NLS, GLM or other}

\item{control}{control parameters (TODO)}
}
\description{
Zhang (2008) model
}
